+++ consumer/build.gradle
@@ -12,6 +12,7 @@ dependencies {
     api("jakarta.annotation:jakarta.annotation-api:$jakarta_annotation_version")
     implementation("org.slf4j:slf4j-api:$slf4j_version")
     implementation("org.jctools:jctools-core:$jctools_version")
+    implementation("org.apache.commons:commons-lang3")
 
     testImplementation(testFixtures(project(':common')))
     testImplementation(testFixtures(project(":spi")))
