+++ consumer/src/main/java/com/flipkart/varadhi/consumer/UnGroupedMessageSrc.java
@@ -1,11 +1,9 @@
 package com.flipkart.varadhi.consumer;
 
-import com.flipkart.varadhi.entities.Message;
 import com.flipkart.varadhi.entities.Offset;
 import com.flipkart.varadhi.spi.services.Consumer;
 import com.flipkart.varadhi.spi.services.PolledMessage;
 import com.flipkart.varadhi.spi.services.PolledMessages;
-import com.google.common.collect.ArrayListMultimap;
 import lombok.RequiredArgsConstructor;
 
 import java.util.concurrent.BlockingQueue;
@@ -32,7 +30,7 @@ public class UnGroupedMessageSrc<O extends Offset> implements MessageSrc {
     private int processPolledMessages(int offset, PolledMessages<O> polledMessages, MessageTracker[] messages) {
         int i = offset;
         for (PolledMessage<O> polledMessage : polledMessages) {
-            MessageTracker messageTracker = new UnGroupedMessageTracker<>(consumer, polledMessage);
+            MessageTracker messageTracker = new PolledMessageTracker<>(consumer, polledMessage);
             if (i < messages.length) {
                 messages[i++] = messageTracker;
             } else {
@@ -48,15 +46,5 @@ public class UnGroupedMessageSrc<O extends Offset> implements MessageSrc {
             messages[i++] = buffer.poll();
         }
         return i;
-    }
-
-    // TODO(aayush): dummy class for poc
-    static class PolledMessageWrapper<O extends Offset> extends Message {
-        PolledMessage<O> polledMessage;
-
-        public PolledMessageWrapper(PolledMessage<O> polledMessage) {
-            super(polledMessage.getPayload(), ArrayListMultimap.create());
-            this.polledMessage = polledMessage;
-        }
     }
 }
