+++ consumer/src/main/java/module-info.java
@@ -4,6 +4,7 @@ module com.flipkart.varadhi.consumer {
     requires java.base;
     requires com.google.common;
     requires io.netty.common;
+    requires org.apache.commons.lang3;
 
     requires com.flipkart.varadhi.common;
     requires com.flipkart.varadhi.entities;
